# swagger/combined.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Documentation",
    "description": "API description in Markdown."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create a new user",
        "description": "Create a new user",
        "operationId": "createUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "example@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "examples": {
              "application/json": {
                "code": 200,
                "message": null,
                "data": {
                  "id": "b17b6135-7b85-49b1-acc4-9204dd350017",
                  "email": "quangnd.nta@gmail.com",
                  "createdAt": "2024-10-08T01:59:08.062Z",
                  "updatedAt": "2024-10-08T01:59:08.062Z"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}/avatar": {
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update user avatar",
        "description": "Upload or update the avatar of a specific user",
        "operationId": "updateUserAvatar",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the user to update the avatar",
            "required": true,
            "type": "string",
            "example": "b17b6135-7b85-49b1-acc4-9204dd350017"
          },
          {
            "name": "avatar",
            "in": "formData",
            "description": "User avatar file",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "examples": {
              "application/json": {
                "code": 200,
                "message": "Avatar updated successfully",
                "data": {
                  "id": "b17b6135-7b85-49b1-acc4-9204dd350017",
                  "avatarUrl": "https://example.com/uploads/avatar_b17b6135.jpg",
                  "updatedAt": "2024-10-08T02:15:08.062Z"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or file format",
            "examples": {
              "application/json": {
                "code": 400,
                "message": "Invalid file format"
              }
            }
          },
          "404": {
            "description": "User not found",
            "examples": {
              "application/json": {
                "code": 404,
                "message": "User not found"
              }
            }
          }
        }
      }
    }
  }
}